# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Flower Shop API
  version: 1.1.0
  description: |-
    API for the FlowerApplication backend. Supports versioned routes under /api/v1.
servers:
  - url: /api/v1
    description: Versioned API
  - url: /api
    description: Legacy alias (same routes as v1)
...
paths:
...
  /flowers:
    get:
      tags: [Flowers]
      summary: List flowers with filtering, sorting, pagination
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Fields"
        - name: q
          in: query
          schema:
            type: string
          description: Search by name or description
        - name: categoryId
          in: query
          schema:
            type: string
        - name: floristId
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Flower list
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Flower"
                  - $ref: "#/components/schemas/PaginatedFlowers"
    post:
      tags: [Flowers]
      summary: Create flower (florist/admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlowerInput"
      responses:
        "201":
          description: Created flower
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flower"
  /flowers/search:
    get:
      tags: [Flowers]
      summary: Search flowers
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  flowers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Flower"
  /flowers/search/advanced:
    get:
      tags: [Flowers]
      summary: Advanced search
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  flowers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Flower"
...
  /flowers/category/{categoryId}:
    get:
      tags: [Flowers]
      summary: List flowers by category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Flower list
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Flower"
                  - $ref: "#/components/schemas/PaginatedFlowers"
  /flowers/city/{city}:
    get:
      tags: [Flowers]
      summary: List flowers by city
      parameters:
        - name: city
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Flower list
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Flower"
                  - $ref: "#/components/schemas/PaginatedFlowers"
  /flowers/{id}:
    get:
      tags: [Flowers]
      summary: Get flower
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Flower
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flower"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Flowers]
      summary: Update flower (florist/admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlowerInput"
      responses:
        "200":
          description: Updated flower
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flower"
    delete:
      tags: [Flowers]
      summary: Delete flower (florist/admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
...
  /orders/flower/{flowerId}:
    get:
      tags: [Orders]
      summary: List orders containing a flower (florist/admin)
      security:
        - bearerAuth: []
      parameters:
        - name: flowerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Order"
                  - $ref: "#/components/schemas/PaginatedOrders"
...
  /payments:
    post:
      tags: [Payments]
      summary: Create payment (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Payment"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/order/{orderId}:
    get:
      tags: [Payments]
      summary: List payments for order (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
  /routes:
    post:
      tags: [Routes]
      summary: Create delivery route (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Route"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/order/{orderId}:
    get:
      tags: [Routes]
      summary: Get delivery route for order (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
...
  /florists/flowers:
    get:
      tags: [Florists]
      summary: Florist flowers (florist)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Flower list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flower"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource id
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
      description: Page number (enables pagination when provided)
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Page size (enables pagination when provided)
    Sort:
      name: sort
      in: query
      schema:
        type: string
      description: Comma-separated fields, prefix with '-' for desc (e.g. createdAt,-price)
    Fields:
      name: fields
      in: query
      schema:
        type: string
      description: Comma-separated fields to include
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
...
    FlowerInput:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        description:
          type: string
        image_url:
          type: string
        available:
          type: boolean
        categoryId:
          type: string
        floristId:
          type: string
        city:
          type: string
    Flower:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        price:
          type: number
        description:
          type: string
        image_url:
          type: string
        available:
          type: boolean
        categoryId:
          type: string
        floristId:
          type: string
        city:
          type: string
        createdAt:
          type: string
          format: date-time
...
    OrderItem:
      type: object
      properties:
        _id:
          type: string
        orderId:
          type: string
        flowerId:
          type: string
        quantity:
          type: integer
        price:
          type: number
...
    Order:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        floristId:
          type: string
        deliverId:
          type: string
        status:
          type: string
        totalPrice:
          type: number
        city:
          type: string
        deliveryAddress:
          type: string
        orderNumber:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
...
    Payment:
      type: object
      properties:
        _id:
          type: string
        orderId:
          type: string
        amount:
          type: number
        method:
          type: string
        status:
          type: string
        paidAt:
          type: string
          format: date-time
    Route:
      type: object
      properties:
        _id:
          type: string
        orderId:
          type: string
        startPoint:
          type: string
        endPoint:
          type: string
...
    PaginatedFlowers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Flower"
        pagination:
          $ref: "#/components/schemas/Pagination"
    PaginatedOrders:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        pagination:
          $ref: "#/components/schemas/Pagination"
...